AUM-API-Docker/requirements.txt
colorama==0.4.4
fastapi==0.115.0
motor==3.6.0
pydantic==2.10.3
python-dotenv==1.0.0
uvicorn==0.24.0.post1
pymongo==4.9.2
python-multipart
